Separate browse into component and remove loading indicator for items
