+++ client/app/components/browse-list/template.hbs
@@ -1,0 +1,112 @@
+<div class="coll-items">
+    <div class="container">
+        <div class="row m-b-lg">
+            {{#if list}}
+                <div class="col-xs-6">
+                    {{#if organizeMode}}
+                        {{#unless groupView}}
+                            <button class="btn btn-default" {{action 'toggleGroupConfirmation'}}>Group Selected</button>
+                        {{/unless}}
+                        <button class="btn btn-danger" {{action 'toggleDeleteConfirmation'}}>Delete selected </button>
+                    {{/if}}
+                </div>
+                <div class="col-xs-6 text-right">
+                    {{#if organizeMode}}
+                        <button class="btn btn-default" {{action 'toggleOrganizeMode'}}>Cancel organizing </button>
+                    {{else}}
+                        <button class="btn btn-info" {{action 'toggleOrganizeMode'}}>Organize Items </button>
+                    {{/if}}
+                    <div class="btn-group">
+                        <button class="btn {{if cardView 'btn-info' 'btn-default'}}" {{action 'changeView' true}}>
+                            <i class="fa fa-th"></i>
+                        </button>
+                        <button class="btn {{unless cardView 'btn-info' 'btn-default'}}" {{action 'changeView' false}}>
+                            <i class="fa fa-list"></i>
+                        </button>
+                    </div>
+                </div>
+            {{/if}}
+        </div>
+        <div class="row">
+            {{#if list}}
+                {{#if cardView}}
+                    {{#each list as |item|}}
+                        {{collection-item item=item cardView=cardView toggleSelectedList=(action 'toggleSelectedList') organizeMode=organizeMode class="col-sm-6 col-lg-4"}}
+                    {{/each}}
+                {{else}}
+                    <table class="coll-table">
+                        <thead>
+                            <tr>
+                                <th>Title</th>
+                                <th>Type</th>
+                                <th>Status</th>
+                                <th></th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            {{#each list as |item|}}
+                                {{collection-item item=item cardView=cardView tagName='tr' toggleSelectedList=(action 'toggleSelectedList') organizeMode=organizeMode}}
+                            {{/each}}
+                        </tbody>
+                    </table>
+                {{/if}}
+            {{else}}
+                <p class="text-center text-muted p-xl"> You don't have any items yet. Go to "Add content" to add items</p>
+            {{/if}}
+        </div>
+    </div>
+</div>
+
+{{#if showDeleteConfirmation}}
+<div class="modal fade in" tabindex="-1" role="dialog">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <button type="button" class="close" aria-label="Close" {{action 'toggleDeleteConfirmation'}}><span aria-hidden="true">&times;</span></button>
+        <h4 class="modal-title">Delete items?</h4>
+      </div>
+      <div class="modal-body">
+        <p>The following items will be removed:</p>
+        <ul>
+            {{#each selectedItems as |item|}}
+                <li>{{item.title}}</li>
+            {{/each}}
+        </ul>
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-default" {{action 'toggleDeleteConfirmation'}}>Close</button>
+        <button type="button" class="btn btn-danger" {{action 'deleteSelected'}}>Delete</button>
+      </div>
+    </div><!-- /.modal-content -->
+  </div><!-- /.modal-dialog -->
+</div><!-- /.modal -->
+<div class="modal-backdrop fade in"></div>
+{{/if}}
+
+{{#if showGroupConfirmation}}
+<div class="modal fade in" tabindex="-1" role="dialog">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <button type="button" class="close" aria-label="Close" {{action 'toggleGroupConfirmation'}}><span aria-hidden="true">&times;</span></button>
+        <h4 class="modal-title">Group selected items?</h4>
+      </div>
+      <div class="modal-body">
+        <p> Enter title for Group </p>
+        {{input class="form-control" value=groupTitle placeholder="e.g. Microbiology studies"}}
+        <p class="m-t-md">The following items will be added to the group:</p>
+        <ul>
+            {{#each selectedItems as |item|}}
+                <li>{{item.title}}</li>
+            {{/each}}
+        </ul>
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-default" {{action 'toggleGroupConfirmation'}}>Close</button>
+        <button type="button" class="btn btn-success {{unless groupTitle 'disabled'}}" {{action 'groupSelected'}}>Group</button>
+      </div>
+    </div><!-- /.modal-content -->
+  </div><!-- /.modal-dialog -->
+</div><!-- /.modal -->
+<div class="modal-backdrop fade in"></div>
+{{/if}}
