+++ client/app/templates/collection/group/index.hbs
@@ -1,77 +1,2 @@
 {{group-detail model=model changeRoute=(action 'changeRoute')}}
-
-<div class="coll-items">
-    <div class="container">
-        <div class="row m-b-lg">
-            {{#if model.items}}
-                <div class="col-xs-6">
-                    {{#if organizeMode}}
-                        <button class="btn btn-danger" {{action toggleProperty 'showDeleteItemConfirmation'}}>Delete selected </button>
-                    {{/if}}
-                </div>
-                <div class="col-xs-6 text-right">
-                    <button class="btn btn-info" {{action toggleProperty 'organizeMode'}}>Organize Items </button>
-                    <div class="btn-group">
-                        <button class="btn {{if cardView 'btn-info' 'btn-default'}}" {{action 'changeView' true}}>
-                            <i class="fa fa-th"></i>
-                        </button>
-                        <button class="btn {{unless cardView 'btn-info' 'btn-default'}}" {{action 'changeView' false}}>
-                            <i class="fa fa-list"></i>
-                        </button>
-                    </div>
-                </div>
-           {{/if}}
-        </div>
-        <div class="row">
-            {{#if model.items}}
-                {{#if cardView}}
-                    {{#each model.items as |item|}}
-                        {{collection-item item=item parent='group' toggleSelectedList=(action 'toggleSelectedList') organizeMode=organizeMode class="col-sm-6 col-lg-4"}}
-                    {{/each}}
-                {{else}}
-                    <table class="coll-table">
-                        <thead>
-                            <tr>
-                                <th>Title</th>
-                                <th>Type</th>
-                                <th>Status</th>
-                                <th></th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {{#each model.items as |item|}}
-                                {{collection-item item=item cardView=cardView tagName='tr' parent='group' toggleSelectedList=(action 'toggleSelectedList') organizeMode=organizeMode}}
-                            {{/each}}
-                        </tbody>
-                    </table>
-                {{/if}}
-            {{/if}}
-        </div>
-    </div>
-</div>
-
-{{#if showDeleteItemConfirmation}}
-<div class="modal fade in" tabindex="-1" role="dialog">
-  <div class="modal-dialog" role="document">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" aria-label="Close" {{action toggleProperty 'showDeleteItemConfirmation'}}><span aria-hidden="true">&times;</span></button>
-        <h4 class="modal-title">Delete items?</h4>
-      </div>
-      <div class="modal-body">
-        <p>The following items will be removed:</p>
-        <ul>
-            {{#each selectedItems as |item|}}
-                <li>{{item.title}}</li>
-            {{/each}}
-        </ul>
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-default" {{action toggleProperty 'showDeleteItemConfirmation'}}>Close</button>
-        <button type="button" class="btn btn-danger" {{action 'deleteSelected'}}>Delete</button>
-      </div>
-    </div><!-- /.modal-content -->
-  </div><!-- /.modal-dialog -->
-</div><!-- /.modal -->
-<div class="modal-backdrop fade in"></div>
-{{/if}}
+{{browse-list model=model changeRoute=(action 'changeRoute') groupView=true}}
