+++ client/tests/integration/components/browse-list/component-test.js
@@ -1,0 +1,25 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('browse-list', 'Integration | Component | browse list', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{browse-list}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#browse-list}}
+      template block text
+    {{/browse-list}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
